requirements.txt
playwright==1.49.1
requests==2.32.3
python-decouple==3.8